requirements.txt
numpy
pydantic
openai
python-dotenv